Fix Windows link and C++20 extension errors uncovered by the warning cleanup
The C4244/C4458 cleanup let MSVC compile further, exposing three latent
Windows-only failures:

- LNK2019 'main' unresolved in every example and game: SDL.h renames
  main->SDL_main, so each executable target must link SDL2main.lib, not
  just sdl_app. Added the SDL2::SDL2main link to add_umbra_example() and
  add_umbra_game().
- C7555 designated initializers in SlimeMold: C++20 feature the C++17
  build rejects on MSVC (GCC/Clang allow it as an extension). Rewrote the
  PanelMetrics aggregate init in declaration order.
- C4244 int->float inside STL instantiations (xmemory/memory) in
  2DViewportToWorld (Point2D emplace_back) and PhysicsReactive
  (PhysicsObject make_unique) - explicit float args at the call sites.

+++ Examples/2DViewportToWorld/main.cpp
@@ -60,7 +60,7 @@ public:
         // Create some world space points (a grid)
         for (int x = -500; x <= 500; x += 100) {
             for (int y = -500; y <= 500; y += 100) {
-                worldPoints.emplace_back(x, y);
+                worldPoints.emplace_back((float)x, (float)y);
             }
         }
     }
